stream: Move orderer processor from node to stream
Here we remove the orderer processor implementation (along with the
associated types) from `p2panda` and replace the existing implementation
in `p2panda-stream` with it. This requires changes to several tests,
since the previous implementation returned an operation but we now
return a tuple of an event and result.
