+++ p2panda-stream/Cargo.toml
@@ -39,7 +39,10 @@ ingest = [
     "dep:serde",
 ]
 log_prune = []
-orderer = []
+orderer = [
+  "p2panda-store/processor",
+  "dep:serde",
+]
 spaces = [
   "p2panda-auth",
   "p2panda-spaces",
