+++ p2panda/src/processor/mod.rs
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
 mod event;
-mod orderer;
 mod pipeline;
 mod tasks;
